Create the state endpoint as a beacon API
