+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v2/debug/GetStateTest.java
@@ -30,6 +30,7 @@ import java.util.concurrent.ExecutionException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.beaconrestapi.AbstractMigratedBeaconHandlerWithChainDataProviderTest;
+import tech.pegasys.teku.beaconrestapi.handlers.v1.beacon.GetState;
 import tech.pegasys.teku.spec.datastructures.metadata.StateAndMetaData;
 
 class GetStateTest extends AbstractMigratedBeaconHandlerWithChainDataProviderTest {
